+++ src/proposals/mips/mip-m24/mip-m24.sol
@@ -180,7 +180,7 @@ contract mipm24 is HybridProposal, MultichainGovernorDeploy {
     }
 
     function run(Addresses addresses, address) public override {
-        vm.selectFork(uint256(primaryForkId()));
+        vm.selectFork(primaryForkId());
 
         _runMoonbeamMultichainGovernor(addresses, address(1000000000));
 
@@ -190,7 +190,7 @@ contract mipm24 is HybridProposal, MultichainGovernorDeploy {
         _runBase(addresses, temporalGovernor);
 
         // switch back to the moonbeam fork so we can run the validations
-        vm.selectFork(uint256(primaryForkId()));
+        vm.selectFork(primaryForkId());
     }
 
     function validate(Addresses addresses, address) public override {
@@ -382,6 +382,6 @@ contract mipm24 is HybridProposal, MultichainGovernorDeploy {
             "MultichainGovernor not trusted sender"
         );
 
-        vm.selectFork(uint256(primaryForkId()));
+        vm.selectFork(primaryForkId());
     }
 }
